+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # RAFT 0.15.0 (Date TBD)
 
 ## New Features
+- PR #12: Spectral clustering.
 - PR #7: Migrating cuml comms -> raft comms_t
 - PR #15: add exception based error handling macros
 - PR #29: Add ceildiv functionality